requirements.txt
playwright==1.44.0
pyyaml==6.0.1
python-dateutil==2.9.0
